Refactor: Unify test drivers with single parseOpenACC API
- Unified all test drivers to use only parseOpenACC() API
- Removed all direct ANTLR4 API calls from test code
- Fixed memory leaks in parseOpenACC() and test drivers
- Restructured build system for external project compatibility

Changes:
- src/OpenACCASTConstructor.cpp: Fix ParseTreeWalker memory leak, remove unsafe assert
- tests/: Move acc_roundtrip.cpp from utils/, remove acc_demo (not needed)
- tests/CMakeLists.txt: Build all test executables locally, only depend on libaccparser
- CMakeLists.txt: Remove hardcoded antlr4-runtime path, add dynamic discovery
- tests/builtin/CMakeLists.txt: Fix CMAKE_SOURCE_DIR paths, add target checks
- Remove utils/ directory entirely - all test code now in tests/

Results:
- All 914 tests pass (42 builtin + 872 OpenACC VV)
- Test executables only link to libaccparser.so
- Works with external projects via add_subdirectory()
- Clear error when accparser target missing
