+++ tests/builtin/CMakeLists.txt
@@ -1,3 +1,7 @@
+if(NOT TARGET acc_tester.out)
+    message(FATAL_ERROR "acc_tester.out target not found. Ensure it is defined in tests/CMakeLists.txt before adding builtin subdirectory.")
+endif()
+
 enable_testing()
 add_custom_target(acc_base_test)
 
@@ -49,7 +53,7 @@ add_dependencies(acc_base_test acc_tester.out)
 foreach(file_to_test ${TEST_FILES})
     add_test(
     NAME ${file_to_test}
-    COMMAND ${CMAKE_SOURCE_DIR}/tests/test.sh ${CMAKE_CURRENT_SOURCE_DIR}/${file_to_test} ${CMAKE_CURRENT_SOURCE_DIR}/reference/ref_${file_to_test}
+    COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/../../tests/test.sh ${CMAKE_CURRENT_SOURCE_DIR}/${file_to_test} ${CMAKE_CURRENT_SOURCE_DIR}/reference/ref_${file_to_test}
     WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/tests
     )
 endforeach()
